+++ src/adapters/TracksAdapter.ts
@@ -40,6 +40,10 @@ export class TracksAdapter implements ITaskAdapter {
 
       return new PingResult(true, true);
     } catch (ex) {
+      if (!ex || typeof ex !== "object" || !("status" in ex) || typeof ex.status !== "number") {
+        return new PingResult(false, false, "An exception occurred during ping. Returned type is unknown.");
+      }
+
       // 401 = auth failed
       if (ex.status === 401) {
         return new PingResult(true, false);
@@ -145,11 +149,12 @@ export class TracksAdapter implements ITaskAdapter {
     return true;
   }
 
-  public async createTask(contextId: number, text: string): Promise<TaskItem> {
+  public async createTask(contextId: number, title: string, description: string): Promise<TaskItem> {
     try {
       const xmlBody = `<todo>
-    <description>${text}</description>
+    <description>${title}</description>
     <context-id>${contextId}</context-id>
+    <notes>${description}</notes>
   </todo>`;
 
       const response = await this.doRequest({
@@ -166,7 +171,7 @@ export class TracksAdapter implements ITaskAdapter {
       const parts = location.split("/");
       const newId = parseInt(parts[parts.length - 1], 10);
 
-      return new TaskItem(newId, text);
+      return new TaskItem(newId, title);
     } catch (e) {
       // todo: handle error
       console.error("error creating task: " + e);